requirements.txt
groq
numpy
sounddevice
soundfile
pynput
python-dotenv